double instead of int. add comment

+++ src/main/java/ArrayBasics.java
@@ -1,14 +1,19 @@
 public class ArrayBasics {
     public static void main (String[] args) {
-        int[] numbers = {1, 2, 3, 4, 10, 6, 7, 8, 9, 5};
+        double[] numbers = {1, 2, 3, 4, 10, 6, 7, 8, 9, 5};
 
         System.out.println(findLargestNumber(numbers));
     }
 
-    public static int findLargestNumber (int[] numbers) {
-        int largest = 0;
+    /**
+     * Finds the largest number in an array of doubles and returns it
+     * @param numbers - array of doubles
+     * @return the largest number
+     */
+    public static double findLargestNumber (double[] numbers) {
+        double largest = 0;
 
-        for(int number : numbers) {
+        for(double number : numbers) {
             if (number > largest) {
                 largest = number;
             }
